# Conflicts:
#	Assets/Project/Materials/Enviroment/OuterPlaneGround_Mat.mat
#	Assets/Project/Materials/Tiles/ObstacleTileTop_Mat.mat
#	Assets/Project/Materials/Tiles/TileTop_Mat.mat
#	Assets/Project/Scenes/TDLevels/LevelsEarly/E2_Valhala.unity

+++ Assets/Project/Scripts/Enemies/EnemyWaves/EnemyWave.cs
@@ -12,7 +12,7 @@ public class EnemyInWave
 
     public EnemyTypeConfig EnemyType => _enemyType;
     public float DelayBeforeSpawn => delayBeforeSpawn;
-    public float NumberOfSpawns => numberOfSpawns;
+    public int NumberOfSpawns => numberOfSpawns;
 
 
     public EnemyInWave()
